Add files via upload
"In the year 2017, India witnessed a total of 1,167 individuals tragically affected by violence, encompassing both injuries and fatalities. These numbers, while distressing, represent a fraction of the global picture. Across the world, 51,372 people experienced the devastating consequences of conflicts and violence in that same year. These statistics serve as a stark reminder of the urgent need for peace, unity, and global cooperation to ensure a safer and more harmonious future for all."